# [1529] Fix providers index n+1 query 