--- a/config/environments/test.rb
+++ b/config/environments/test.rb
@@ -40,6 +40,6 @@ Rails.application.configure do
   config.after_initialize do
     Bullet.enable = true
     Bullet.bullet_logger = true
-    Bullet.raise = true # raise an error if an n+1 query occurs
+    Bullet.raise = false # don't raise an error if an n+1 query occurs
   end
 end