Publish users can't currently load the page.

### Benchmarks

Before:

```
Completed 200 OK in 13733ms (Views: 12721.4ms | ActiveRecord: 985.5ms)
Completed 200 OK in 13739ms (Views: 12755.9ms | ActiveRecord: 981.3ms)
Completed 200 OK in 13950ms (Views: 12982.5ms | ActiveRecord: 965.6ms)
Completed 200 OK in 15763ms (Views: 14690.1ms | ActiveRecord: 1071.0ms)
Completed 200 OK in 14026ms (Views: 13031.5ms | ActiveRecord: 993.1ms)
```

After:

```
Completed 200 OK in 658ms (Views: 582.9ms | ActiveRecord: 48.9ms)
Completed 200 OK in 877ms (Views: 834.7ms | ActiveRecord: 40.8ms)
Completed 200 OK in 580ms (Views: 537.7ms | ActiveRecord: 41.2ms)
Completed 200 OK in 504ms (Views: 459.9ms | ActiveRecord: 42.0ms)
Completed 200 OK in 913ms (Views: 869.3ms | ActiveRecord: 42.5ms)
```

### Checklist

- [x] Make sure all information from the Trello card is in here
- [x] Attach to Trello card
- [x] Rebased master
- [x] Cleaned commit history
- [x] Tested by running locally