@tvararu 2019-05-24
![](https://github.trello.services/images/mini-trello-icon.png) [Fix n+1 queries](https://trello.com/c/gjJXup36/1529-fix-n1-queries)